moved images to a different folder
